use bars instead of monitors
We are moving away from the concept of each monitor necessarily having
its own bar and instead adopting a model that accepts arbitrary bars in
any location. As a result it no longer makes sense to refer to a bar
container as a Monitor so we are migrating from Monitor to Bar.

TODO: Currently the Bar class is also managing its window. This should
be handled by a seperate Window class which Bar will have an instance
of.